Ed tidying up his changes

+++ IDSort/src/v2/magnet_tools.py
@@ -97,8 +97,8 @@ def calculate_phase_error(info, b_array):
     
     nperiods=info['periods']
     step=info['sstep']
-    n_stp = (info['period_length']/step)
-    n_s_stp = round(((info['smax']-info['smin'])/step))
+    n_stp = int(info['period_length']/step)
+    n_s_stp = int(round(((info['smax']-info['smin'])/step)))
     
 #    logging.debug("step %s"%(str(info['sstep'])))
 #    logging.debug("smax %s"%(str(info['smax'])))
